pass theme in extra file
