+++ ui/src/App.tsx
@@ -7,8 +7,8 @@ import { RoutePath } from "./interfaces/router.interface";
 import { TranscendanceContext } from "./context/transcendance-context";
 import { TranscendanceStateActionType } from "./context/transcendance-reducer";
 import Toast from "./context/toast";
-import { Box, createTheme, ThemeProvider } from "@mui/material";
-import classes from "./styles.module.scss";
+import { Box, ThemeProvider } from "@mui/material";
+import theme from "./theme";
 
 export default function App() {
   function closeToast() {
@@ -20,75 +20,6 @@ export default function App() {
 
   const { transcendanceState, dispatchTranscendanceState } =
     React.useContext(TranscendanceContext);
-
-  const theme = createTheme({
-    palette: {
-      primary: {
-        main: classes.colorPrimary,
-        light: classes.colorSecondary,
-      },
-      secondary: {
-        main: classes.colorAccent,
-      },
-    },
-    typography: {
-      fontFamily: "Arial",
-    },
-    components: {
-      MuiInputLabel: {
-        styleOverrides: {
-          root: {
-            fontSize: "14px",
-            fontWeight: "500",
-          },
-        },
-      },
-      MuiOutlinedInput: {
-        styleOverrides: {
-          root: {
-            fontSize: "14px",
-            fontWeight: "400",
-            borderRadius: "2px",
-          },
-        },
-      },
-      MuiFormLabel: {
-        styleOverrides: {
-          root: {
-            fontSize: "11px",
-            fontWeight: "500",
-          },
-        },
-      },
-      MuiInputBase: {
-        styleOverrides: {
-          root: {
-            backgroundColor: "white",
-            "&.Mui-focused": {
-              color: classes.colorPrimary,
-            },
-            fontSize: "14px",
-            fontWeight: "400",
-            height: "56px",
-          },
-        },
-      },
-      MuiAutocomplete: {
-        styleOverrides: {
-          root: {
-            height: "56px",
-            backgroundColor: "white",
-          },
-          listbox: {
-            "& .MuiAutocomplete-option": {
-              fontSize: "14px",
-              fontWeight: "400",
-            },
-          },
-        },
-      },
-    },
-  });
 
   return (
     <>
